moved tests to tests.py and cleaned tests
